+++ src/plugins/replication-graphql/index.ts
@@ -52,7 +52,10 @@ import type {
     RxDocumentData
 } from '../../types';
 import { getDocumentDataOfRxChangeEvent } from '../../rx-change-event';
-import { _handleFromStorageInstance, _handleToStorageInstance } from '../../rx-collection-helper';
+import {
+    _handleFromStorageInstance,
+    _handleToStorageInstance
+} from '../../rx-collection-helper';
 
 addRxPlugin(RxDBLeaderElectionPlugin);