+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Bugfixes:
 
 Other:
   - We now have set `sideEffects: false` to the default in the package.json so tree shaking can work.
+  - Optimized memory usage in the query cache.
 
 
 ### 10.1.0 (27 September 2021)